requirements.txt
asgiref==3.7.2
Django==4.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
mysqlclient==2.1.1
Pillow==9.5.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3

mkdocs
mkdocstrings[python]
mkdocs-material
